Update the logic on the results page
We use `tier_three?` and `tier_two?` to determine which results page
to show.

At the moment we need to the same results page to show regardless of whether
there are current or future restrictions. That's because at the moment an area
only has one or the other, so we wanted to make sure that the same results page
displayed on the 1st and 2nd of December when the new tiers come into effect.

From 2 December this changes, and all areas will have a "current" restriction, so
we need to put the logic back to the way it was before, only looking at the current
tier level.