Make the postcode checker heading line break in the right place
The results of the postcode checker have the heading 'This area is in [full tier name]'.

We want to make a line break so the full tier name is always on one line so people can easily understand it. So the options will display as:

This area is in
Tier 1: Medium alert

This area is in
Tier 2: High alert

This area is in
Tier 3: Very High alert

So people can easily identify their tier

https://trello.com/c/LCDIMFk5/967-make-the-postcode-checker-heading-line-break-in-the-right-place