Restructure/refactor LocalRestriction model
This changes the interface of LocalRestriction to be closer fit the
model design pattern, particularly for a Rails application. It has a
constructor that now takes the attributes of a model and a static
interface to look items up (similar to ActiveRecord).

This change now stores all LocalRestrictions in memory on the first call
for restrictions, this saves every request involving a read of the YAML
file.

This refactors the tests to reflect these changes.