Fix leaky locale assignment
A number of controllers in this app were using a before_action to set
the value of `I18n.locale` based on the request locale. This changes the
locale for the entire application, including subsequent requests. [1]

We learnt of this becoming a problem when we heard reports that users of
"/find-coronavirus-local-restrictions" were occasionally seeing a
back button in Welsh (Yn ôl). Investigating further we learnt that this
problem could be replicated by visiting a page in Welsh in this app (such as
[2]) and then visiting the postcode checker. The reason it was only
the back button that exhibited this problem is because it's the only
content with an available Welsh translation on this page.

To resolve this we need to change the approach used to change locale to
be encapsulated to only the particular action that is running. The Rails
documentation explains how this is done [1].

I've taken this as an opportunity to DRY up the locale setting so we
have one action on the superclass that other controllers can use.

[1]: https://guides.rubyonrails.org/i18n.html#managing-the-locale-across-requests
[2]: https://www.gov.uk/government/organisations/office-of-the-secretary-of-state-for-wales.cy