Refactor the local restriction integration tests for changed stub
As LocalRestriction now has a new interface these tests needed updating
for a different stubbing approach. I've taken the need to add this
in as an opportunity to DRY up aspects of these tests.

The main way I've approached this is by using a stub method that can
stub mapit and LocalRestriction from the same call.

I found myself rather confused by areas set in test conditions that were
otherwise unrelated to the scenario (for example
`then_i_see_the_results_page_for_level_one`), so I instead switched to
using an instance variable so that the variable could be defined when
the stub is created.

As this was the last file that referenced the local-restrictions.yaml
test data we can now remove that file.