Create a class to model the postcode search
This is intended to encapsulate the logic that is currently dealt with
in the CoronavirusLocalRestrictionController and PostcodeService. Moving
this into a class allows us to simplify the number of outcomes the view
can have.

The postcode regular expression is taken from the PostcodeService class.
I've switched it to be a multi-line one so it's not one massive line of
code and the parts are understandable.

In the subsequent commits I'm expecting to complicate the controller
further by handling search requests in the same action as a show. This
class will help that process by shortening the amount of logic needed.

The abstraction here isn't quite ideal, it somewhat models a search and
a result together. I've got a plan to resolve this in a subsequent PR.