Handle postcode restrictions via GET requests
This changes the HTTP process for looking up local restrictions by
postcode. Previously there was a GET request to show the form and a POST
request to see results, both with different actions. This changes the
handling to be done with GET requests for both form and results and this
done with a single action.

There is a risk that this change could inconvenience any users who are
in the process of checking their postcode at the time of this change. To
handle these users there is a redirect set-up for any post requests.
This can be removed an hour or so after this has been deployed.

This also change simplifies the controller by switching to the
PostcodeLocalRestrictionSearch model, this allows the conditionals in
the controller to only be the different views. It also moves a bunch of
the phrase handling into the view from the controller.