Remove the shared postcode partial
Nothing else uses this partial so I assume it was added in shared when
there were different expectations about usage. At this point in time it
seems rather coupled to the show template so it seems simpler to move it
into that template.

As part of this change I changed the unconventional camel case id of
postcodeLookup to a simpler postcode id. This should have no effect of
users as this anchor is mostly used by JS and should only apply based on
when the page is rendered.