Move meta tags into dedicated partial
This is the first step towards sharing this with the other views that
can be shown for this endpoint.

I've added in the "govuk:static-analytics:strip-postcodes" meta tag so
that postcodes will be stripped from analytics automatically. I've also
modified the means to specify the canonical URL so that it's more
conventional.