Set meta tags on results and no_information views
These meta tags will inform Google, Facebook and the wider world that the
canonical resource for this page is the URL without the query string. This
should hopefully help postcode variants not be too widespread.

It will also include the meta attribute to strip postcodes from being
sent to Google Analytics.