Equalize section padding
The `landing-page__section--light-grey` modifier is currently being used in two places, the transition landing page and the DIT landing page. Previously we relied on the start button on the transition landing page to have a margin bottom set, which, by adding up with a scale 4 padding-bottom would result is roughly the same as padding-top. We now set equal padding (top and bottom) and removed the margin-bottom from the start button on the transition landing page to make sure the applied padding is the same on any section using this modifier, regardless of the content.