Return cache headers for Coronavirus local restrictions
This returns a default cache header of 30 minutes public for all
responses except when there is an upcoming future restriction.

By default Rails will return a cache header of max-age=0, public=false.
Somewhere in the GOV.UK stack this seems to get changes to a public 30
minute cache implicitly. This feels a bit dubious as we don't know what
part of the stack is controlling the cache so it seems better to set
this explicitly.

When a future restriction is coming we don't want users to be served a
page telling them the wrong information. When the time until a new
restriction is less than 30 minutes we don't want to tell users they can
cache the page for 30 mins, instead we return the time until the
restriction takes effect. So for a future restriction starting in 5
minutes, we tell users they can cache the page for 5 minutes.