Add a 3 hour Rails cache to mapit postcode lookup
This line returns the area a postcode is in. Area boundaries are
infrequently updated, (rarely more than twice a year).

Repeated queries for the same postcode will return the same area, making
this a good candidate for caching to boost performance.

We use a three hour expiry because long enough that it should help cover
spikes in traffic, but not require cache busting when there is a rare
area boundary update.

Co-authored-by: Jon Hallam <jonathan.hallam@digital.cabinet-office.gov.uk>