Flip the partials for christmas rules and future restrictions
All restrictions are meant to be in date order. The new restrictions that come into being before the christmas rule changes are in the wrong place. This swaps the christmas rules and future restrictions so it displays properly. This will have to be reverted before any new restrictions that come in after christmas are made live.