Add an integration test for the postcode normalisation
It appears that we have a convention for integration testing many of the
scenarios for the postcode form, thus I felt it was relevant to add one
that covered the JS enhancement which normalises the postcode.

Adding this in was something of a pain as it turns out this app wasn't
configured with the dependencies needed for JS testing so I had to add
in the test-dependencies.js file.

Aside from this, the test is rather straight forward and tries to not
cover the same assertions of other integration tests too vividly.