Display 'No information' page if restriction is not defined
- Previously we relied on 'else' logic to display the tier 1 information and did not have
tier 1 restrictions defined in local-restrictions.yml

- This worked fine given that our data was accurate, but if we were in a situation
where the postcode was valid but we didn't have restriction information set correctly
then the logic could fall through to incorrectly displaying tier 1 results.

- When displaying results, it now explicitly checks and displays tier 1 results if an
area is in tier 1.

- If an area is not a devolved nation or in tiers 1,2 or 3 then the 'No Information' page is displayed.

- Previously the 'No information' page displayed if no data was returned by Mapit. The user will
now also see this page if the postcode is valid but we don't have any restriction data in
local_restrictions.yml... this will mean the user is given useful
information in the unlikely case that there is something wrong with our data and a restriction is missing.