Point Scotland link to Scottish checker results
Points the link to the guidance on the Scotland results page to the results page
for the equivalent postcode checker on the Scottish government site.

The whitespace has to be removed from the postcode, otherwise the user will be
directed to the Scottish lookup page and shown an error.

This doesn't affect the link to the Scottish checker lookup page on the lookup page
for this checker as the nil postcode is ignored.