Handle analytics events in postcode form
This ports the code to handle Google Analytics events from the
CoronavirusTrackLocalRestriction module into the
CoronavirusLocalRestrictionsPostcodeForm module. This will allow us to
remove the tracking module and remove fragmentation.

These analytic events are ported as they were, as I didn't want to break
anything. I did wonder whether we have much value for recording an event
on each page view - since we have an implicit one from a page view. I
was a bit perplexed by the namespacing of the events.