Rename LocalRestriction to CoronavirusRestrictionArea
This change was made because the LocalRestriction
model does not actually represent the concept of a restriction and
instead represents an area that may have restrictions.

I chose the name CoronavirusRestrictionArea as these are areas and the
only determining characterising of the area seemed to be that are used
to represent the boundaries of a place that may have Coronavirus
restrictions. I did consider calling it CoronavirusArea but I was
worried that this made each area sound like it was a Coronavirus
hotspot.

I'm not super pleased with this name so other suggestions welcome.
However I feel like this is an easier to understand abstraction.