Simplify CoronavirusRestrictionArea model
This adds a nested class to CoronavirusRestrictionArea to represent an
individual Restriction. The use of this allows simplifying both
the #current and #future methods.

I've removed the #current_alert_level, #future_alert_level methods as I
expect the interaction with these to be done through a restriction
object rather than this area object. They don't need replacements as
alert_level is available.

The `tier_*?` methods have been moved onto the Restriction object so
they can be used consistently for both current and future alerts.