Rename MapitPostcodeResponse to MapitLocation
This naming seems to fit better with the nomenclature that is used for
Mapit. We search for the data using a `.location_for_postcode` method on
Mapit we seem to describe the usage of this data in app as a location.

Within the model I've switched from location to referencing area as this
seems to match the Mapit terminology too where this class is initialised
using the area attribute of a response.