Add method to find locations with MapitLocation
This adds a static method that can query mapit to find locations for a
postcode. This has been done to simplify the MapitLocation abstraction
so it doesn't need to be split between this class and
LocationLookupService.

I've created exception classes for the two Mapit error situations that
this application caters for: No location for the postcode; Postcode
given was invalid. Other responses will raise an error. This somewhat
works around GdsApiAdapters lacking a particular exception for a
HTTPBadRequest which seems to be a strange omission (I'll try add this
so I can refactor this code). As we don't do anything with other Mapit
errors (which doesn't seem quite ideal) I didn't think to add a more
generic MapitError exception.

I couldn't see a particular reason why we converted the Mapit Response
into a hash so I converted this to just work with the object that the
GDS API Adapters responds with. To finish this off I simplified some of
the method names and removed some of the tests for code that just
delegates.