Reflect changed modelling in integration test
This updates these integration test to make use of the new
CoronavirusRestrictionArea stubbing. It also consolidates the stubbing
into a single method that can handle current_alert_level being nil or
not.

It also does a little re-ordering, to try make this long file a little
more logical to navigate.