Require a current restriction for CoronavirusRestrictionArea
Our user interface expects every area that we have in the England to
have a current restriction in place, this commit adds an exception that
will blow up if an area is initialised without one.

Doing this allows us to be sure that current_restriction will also
return something so we can simplify code that considers whether this
value returns nil or not.

This change follows on from the work in https://github.com/alphagov/collections/pull/2163
where we introduced tier 1 areas to our list, thereby no longer treating
an area we didn't know as a tier 1 area.