Test and simplify the postcode pattern
It turned out that this pattern wasn't matching the infamously weird GIR
0AA postcode because of the unescaped white space. Whilst fixing this it
became apparent that this postcode does not produce a result for users
anyway as it is not a geographical postcode.

This commit thus simplifies the regex to no longer consider that
postcode and further simplifies it by using the i flag to indicate
case-insensitivity - allowing us to remove the mixed case usage.

I also added some direct tests for the regex.