Show Christmas rules in tiers where they apply
This is dependent on the guidance at
https://www.gov.uk/guidance/guidance-for-the-christmas-period being
updated to reflect the new Christmas guidance.

The behaviour I think we want is:

* If your postcode is in tier 1 and isn't moving to tier 4, the
  Christmas guidance should appear for you
* If your postcode is in tier 2 and isn't moving to tier 4, the
  Christmas guidance should appear for you
* If your postcode is in tier 3 and isn't moving to tier 4, the
  Christmas guidance should appear for you
* If your postcode is in tier 4, the Christmas guidance should NOT
  appear for you
* If your postcode is moving to tier 4, the Christmas guidance should
  NOT appear for you

I haven't tried to be any more clever than that (e.g. handling areas
which are moving to tier 4 _after_ Christmas), because I think it's
unlikely these situations will arise in the next 6 days.