Fix bug for tier 1 postcodes
Over the weekend we introduced some code to conditionally display christmas rules. Whilst this worked for the overwhelming majority of places, those areas in tier 1 don't technically have a restriction according to the code. This means that for those users things won't have displayed. This adds an extra check for the presence of a restriction before checking the future level of it which fixes the issue.