Stop showing Christmas rules after 25th December
At the moment the Christmas rules only apply to 25 December, so we
should stop showing the rules after that date. However the 25th is a
Friday, and the next working day is the 29th December. That means we
would be showing out of date Christmas rules on the checker for 4 days.

This change uses the date to hide the Christmas rules after 25 December.

The built-in `after?` method is used because that returns true for the
26th December onwards, whereas the `before?` method only returns true
for everything on 24 December and before, not 25 December.

The integration tests that look for Christmas content have also been
consolidated to make them easier to find and remove later on.