Add partial that displays "an out of date warning"
Sometimes restrictions are announced before we have had a chance to
update the postcode checker. It can be hours before the checker is
updated with the latest restriction data.

Users coming to the checker will expect to see up to date information.
We need a way to let them know that their restriction may be changing
and that they should come back later.

A flag on the LocalRestriction model is being used to set whether or not
the data is up to date. This flag is stopgap measure so that the warning
can be displayed in time for the latest set of updates. The upside to
this approach is that it is fast to implement. The downside is that the
application needs to be deployed every time the flag changes.

The long term plan is to use a rake task to set a value in Redis that
determines when the warning should be shown in a similar manner to the
emergency banner.