Adjust the cache time while the service is being updated
When we're updating the service we shouldn't be returning a 30 minute
cache time otherwise users will have a rather odd staggered experience
where the /find-coronavirus-local-restrictions start page shows a
warning long after most of the results are showing updated data.

Setting this to 5 minutes should offer a reasonable balance of not
having our servers hit too frequently during the change, with only a
modest delay to wait for the change to take effect.