Simplify header containers
Since we want all content in the page header to be presented in a 2/3 container on a white background we removed the multiple grid-row definitions and grouped all content into one row in a 2/3 column. This change also aligns the page_header template of the /transition landing page with the one for the /eubusiness landing page with the hope in further iterations to converge to a single 'landing page' template.