Update page header title and content
We don't have a highlighted section anymore so we can remove the logic and styling around it. We keep to govspeak component in case we'd like to render links within the header explainer and to keep it aligned with the /eubusiness template hoping to converge to a single 'landing page' template.