Refactor PostcodeLocalRestrictionSearch to have result objects
This abstracts the result modelling of PostcodeLocalRestrictionSearch to
be done in two separate (and rather simple) classes. One of these
represents a search result for a devolved nation - where we're concerned
about area name and the country, the other represents a search result
for England where we're mostly using data from our restrictions YAML
file. These models are intended to be used for logic regarding a users
result and thus will be mostly used in the view.

To make this class simpler to follow I moved the lazy validation aspects
to be done as part of the initialize method. This sets up everything
from the start, instead of having to follow various methods that could
do it lazily.

I've also removed the no_restriction? method as this won't be needed in
the controller when it is refactored to use this class and I didn't want
to alter it to still work only to remove it in a couple of commits.

---

An area of complexity in this class is how we create the EnglandResult
object. We only want to have an england_result if we have matched the
area to a LocalRestriction, because of this we do this lookup
before we create the class. A previous iteration did this inside the
EnglandResult class however this wasn't great because if a restriction
didn't exist this was something of a null object.

The inspiration for making this change is that the concept of a Search
was rather confused by it representing both a successful state and an
error state. Some aspects meant different things in different contexts,
for example "postcode" may mean the user's input in a search, whereas in
a result we'd expect the postcode to be the postcode with formatting
cleaned up. Breaking this up into separate classes helped remove these
concerns as we the classes themselves represent the context.