Add a future_start_time method to LocalRestriction
This is to make it easier to interact with the start_date and start_time
fields present on a future restriction. It allows the caller to have a
Time object set to the start_date and start_time.

This has been done as a light touch means to make this method available.
It'd arguably be more logical that restrictions were represented by
objects that know the start time and alert_level implicitly rather than
being accessed through this a LocalRestriction object. However this
approach has been taken to minimise the changes that would be needed to
achieve that.