Split results template into england_result and devolved_nation_result
Now that there are different result objects for England and the devolved
nation we can split the result view to have dedicated views for each.
This allows a simplification of the concerns in the view by them only
needing to interact with a result object for data rather than needing to
navigate location_lookup, search and postcodes.

In the controller I've changed the ordering of the conditional for
deciding the view to fallback to no_information rather than it being one
of the earlier cases. This meant we no longer need a no_restriction or
no_information conditional.

The use of result objects allowed various simplifications to the view
logic. One notable change is the removal of `#tier_one?` methods, this
was removed for consistency with future_alert_level methods as it seemed
odd to use them for only one scenario.