Strip all postcode parameters from being sent to Analytics
This depends on: x and can't be merged until that is released.

This changes our approach to stripping postcodes out of strings that are
sent to Google Analytics. Rather than matching a postcode pattern we
instead match any instances of a query string that has a postcode=
parameter and strip the value of that parameter.

This has been done as we've had occurrences of users entering PII that
is not a postcode and, as it doesn't match our postcode pattern, it is not
stripped. This instead strips any user input to the postcode field.

Difference

Given a URL of https://www.integration.publishing.service.gov.uk/find-coronavirus-local-restrictions?postcode=52+made+up+address+London+E1+8QS

Before this change this would be submitted to Analytics as: https://www.integration.publishing.service.gov.uk/find-coronavirus-local-restrictions?postcode=52+made+up+address+London+[postcode]
After this change this would be submitted to Analytics as: https://www.integration.publishing.service.gov.uk/find-coronavirus-local-restrictions?postcode=[postcode]