Preserve original link styles when video partial is used with inverse flag
The youtube video player partial has an --inverse modifier meant to be used when the component is called within an element with a dark background, as it's the case with the transition page. However, in such scenario, we don't protect the links within the component from getting the 'white' rules on links – required for anchors used on a dark background – resulting in invisible links when cookies are not yet accepted or disabled. With this change, we make sure the default link styles are being used when the video player is called with the --inverse flag.