Cache coronavirus restrictions based on cache clear time
This adjust any cache headers we send for the
/find-coronavirus-local-restriction endpoint to be reduced gradually
when a cache clear time is near. This is to ensure that users all
receive the same response at that time

I've chosen to do this for all the pages rather than just result pages
as I'm expecting even the index page could have time-sensitive content.