Configure /find-coronavirus-local-restrictions redirect in controller
We noticed that this redirect isn't cached and a fresh origin request is
made for every request. In order to add caching and to make this be
testable I moved this concern to the controller. I used the same cache
method as is used in the controller so that this can consider whether
out of date is set or not for a shorter cache time.

As we expect to change this redirect back to the destination page I've
marked this redirect with a temporary redirect (307) status code.

This is ported with the same behaviour as the previous constraint of
basing this condition on Rails.env.production?