Fix issues caused by i18n upgrade from v1.8.5 to v1.8.7
A change was introduced in v1.8.6 to support a change in behaviour
in Ruby 3: https://github.com/ruby-i18n/i18n/pull/545

> In Ruby 3.0, the behavior of `I18n.t("key", { value: "foo" })` has
> silently changed to `{ value: "foo" }` is no longer regarded as
> `**options` but regarded as noop slpat arguments `*`.

As such, we have to change our calls to `I18n.t` to take 2 args
rather than 3. This immediately causes an issue whereby if you
pass a count of 2 and no pluralization data, it raises a
`I18n::InvalidPluralizationData`, so we've had to catch that and
set the response to `nil`, which appears to be the behaviour we
were expecting prior.