Fix filter whitespace bug; add filter by acronym
Some of the names in the organisation list have newlines instead of spaces, and
some have double spaces instead of single spaces. Whilst this should also be
fixed at a content level, it's also good for the filter to able to cope with
this.

When an organisation name had a newline or multiple spaces in it the filter was
unable to find it. This is because `Ministry ` (with a space at the end) is not
equivalent to `Ministry`; and `Ministry ` (with a space) is not equivalent to
`Ministry\n`.

This trims the whitespace of the string that's being searched for and normalises
the strings being searched in.

This means that ` department of `, `department of` and `department\nof` are all
treated as equivalent.

An extra `data-filter-acronym` attribute has been added to allow the filter to
filter by organisational acronym - so `MOD` will show `Ministry of Defence`.

Tests have been added and updated to cope with these scenarios.