Update error message content
The error messages are being reconfigured so that regardless of whether
the user has entered nothing for the postcode or the postcode is not in
the correct format, they will see the "invalid_postcode" postcode error.
The "no_postcode" error will only show if mapit returns a 404 for the
postcode.
The first step in this change is updating the content.

N.B. If mapit returns no information at all, the user is directed to the
"no information" page.