Remove blank postcode check
The error messages are being reconfigured so that regardless of whether
the user has entered nothing for the postcode or the postcode is not in
the correct format, they will see the "invalid_postcode" postcode error.
The "no_postcode" error will only show if mapit returns a 404 for the
postcode.

This means we no longer need a separate check for a "blank postcode".

`invalid_postcode?` is checking each error code rather than calling
`!location_lookup.postcode_not_found` or
`!%w[fullPostcodeNoMapitMatch].include?(error_code)` because both of
these require a call to Mapit, whereas checking each error code allows
the local validation to fail first.

We will need to check whether or not the postcode exists, and a new
"postcode not found" check will be added in the next commit.