Add layout header component styles for /transition
We have made some adjustments to use the layout header component in
static where the account navigation should be present. This would be on
the transition page as well as Brexit checker pages.

This is fine, however some of the layout styles are overriding the
layout header component styles in static, resulting in extra padding
around the layout header content. Importing these styles into
collections seems to fix this issue.