Add an optional summary field to Coronavirus timeline
This field has been added to meet an immediate need for contextual text
immediately below the timeline heading and before the timeline entries.

This is being added in a hurry as copy had been signed off that didn't
fully match what was available for editing. We plan to review this once
things are a little quieter.