Add support for the new session header & dev cookie
The user is treated as logged in if:

- The `GOVUK-Account-Session` request header is set
- The `govuk_account_session` cookie is present and `Rails.env.development?`
- The user is in the `LoggedIn` A/B bucket

The A/B approach will be removed when we fully transition to the new
cookie.

We don't validate that the header or cookie actually contains correct
session data.  This is fine because if an attacker supplies a fake
value, all they'll do is change the "sign in" link in the header to a
"sign out" link.