Replace incorrect environment variable name.
GDS_API_ADAPTERS_PACT_VERSION is set in the pact.rake task,
and used by the pact_helper to compose the pact_uri. There was
a typo, and the pact_helper was trying to fetch GDS_API_PACT_VERSION.
As a consequence, pact files for each new build of gds-api-adaptors
have not been validated against the provider, instead the master version
of the pactfile was used. This is of little consequence luckily,
as none of the consumer pact tests have changed since this typo was deployed.