Move non-English and non-Welsh language names to their own locale
We currently have non-English and non-Welsh language names in the
English and Welsh locale files, which are used to display language names
for translated content. This is very bad practice as it means we have to
maintain two copies of the same translations.

Here we update to include the language name in the correct language
locale.