Service to fetch coronavirus.data.gov.uk statistics
This is a class that will load data from coronavirus.data.gov.uk, parse
the response and return a Struct class with attributes of the fields we
intend to show in the view.

The data we intend to show is the amount of people in the UK who have
been vaccinated cumulatively, the number of hospital admissions for the
previous and the number of positive tests for the previous day. The API
returns an array of data fields by date.

It is very common for the data in the API to not be consistent in which
day it is up-to-date for. For example, as per the unit test, we may
have positive tests for the current day, cumulative vaccinations for 2
days ago and hospital admissions for earlier in the week. The result
class is modelled to consider these differences.

This class is intended to be called during the request for /coronavirus
page. I've coded this in quite a defensive way to try to minimise the
risk of problems on coronavirus.data.gov.uk affecting www.gov.uk. It
uses some broad exception handling to minimise an error in the loading
and parsing taking out the whole page.

In my testing I have found the API to be somewhat inconsistent in
performance (varying from ~150ms to 9s for uncached requests) because of
this I've wrote the code to somewhat anticipate this (and hope it
performs well). I've set this app to give up if we don't get a response
in 2 seconds. There is also an eager read to the cache before trying to
load the data from the API. This means that a user that experiences a
timed out request to coronavirus.data.gov.uk can still likely be served
the statistics by using the stale data.

This stale data is available until the race race_condition_ttl expires,
which would be 5 minutes.

I added some conditions so that we'd only notify Sentry when an errors
that aren't typical network blips occur. This was a bit fiddlier than
I'd have liked. I found it somewhat difficult to test the exact scenario
where Faraday raises Faraday::TimeoutError (webmock's .to_timeout tends
to cause a ConnectionFailed exception) so this isn't fully tested. It
could be done with a lambda and a sleep but I don't want to add a sleep
to the tests.

Oh - I put this test in as an RSpec as I noticed there is a PR to port
all the service tests to RSpec: https://github.com/alphagov/collections/pull/2321