Fix services/search specs
We previously converted them to rspec using the automated tool, but they didn't
work. This commit fixes them.

I've taken the opportunity to tweak the SuperGroups class to accept an org slug
rather than a whole organisation. It only passes on the slug anyway, and this
cleans up the spec setup a little too.