Use coronavirus statistics in landing page controller
This loads the statistics in so they can be used in the view. It's
anticipated that there may be times where we're showing the page without
statistics because they failed to load. In these instances we will
reduce the time the page is cached for - to reduce the number of people
served it. We also log it so that we can use Kibana to analyse how often
this problem occurs.