Create NHS Notice partial for April 2021 COVID guidance
This creates a partial that is tailored to display the NHS guidance that
is expected to be agreed for April 2021.

The copy is so far removed from the previous component (header-notice)
that I've decided it is best to create something dedicated for this.

I've gone with an approach that sets up a modestly flexible section
approach that can be used to set the content for this box mostly through
markdown. I did initially see if the whole block could be markdown but
this causes a bunch of spacing issues with the headers (and creates some
unnecessary anchors) it seemed a pretty messy thing to resolve.

I'm somewhat skeptical that this content designer editable structure
will be flexible enough for future needs but I've tried. I've left the
Content Item data hardcoded into the view for now, until it can be
ported to the YAML file.