Adapt /coronavirus statistics for YAML content
This adapts the hard coding of content on the /coronavirus statistics
page to instead use data that is in the content item. This has been done to
be consistent with this existing convention of nearly all content
existing in the content item.

I chose not to include content that formed parts of a sentence, such as
"Last updated $date" and "$statistic million", as it seemed impractical
to edit those outside of the context of the variable they referenced.

As part of this change I updated some of the variable names for clearer
meaning (stat => statistic, subtext => statistic_meaning) and some for
consistency (href => url).