Remove screenshots
These take up too much space for the little value they add, noting
that each one is also represented in the live examples section. If
we need to know what the up-to-date version of a page looks like, we
can either visit it on GOV.UK, or start this app locally.

In general, we should minimise the length of the README and prioritise
actionable instructions, so that devs can see things for themsevles.