Add fixture content for Coronavirus hub page related links
This change adds fixture content for the Coronavirus hub page - related links component.

There does not appear to be any tests for the existing topic component - most likely because of the way the content for this is managed - by a content designer via GitHub repo (govuk-coronavirus-content) - and the frequency of these changes. However, there is fixture content available for other similar components (e.g. the topic section).

Because of this we have added fixture content for the new hub page related links components, but have not added any additional tests for them.

We note that fixture content does not exist for the worker-support hub page. So, would suggest this is added as part of a different story.