Update browse page
Some inherited styles from govuk_layout needed adding into the browse
stylesheet to ensure that the layout wouldn't change.

Fixed a slight visual discrepancy where a focused link in the browse that was
hovered still had a small yellow bar at the top of the link. This has been
changed to grey to match the hover state.