Remove video player section template & locales
This was commented out in https://github.com/alphagov/collections/pull/2207
on the basis it may need to come back soon - it's been four months
now, so I think is safe to remove.

I did consider leaving an 'empty' locale file for
coronavirus_landing_page.yml but as we're trying to standardise
our locale files, deleting it seems to the cleanest option.