Fix Show more organisations link
The "Show more organisations" link at the bottom of the Organisations
section on the /corporate-information page does not work as intended.
The link should function as a toggle where clicking it shows the full
list of organisations. This changes the data-module attribute on this
partial from "toggle" to "gem-toggle", which restores that
functionality.
Additionally, apply the correct govuk-link class so that the focus state
is in line with the rest of govuk.