Remove live stream feature tests
As "ask a question" is part of the live stream section, so these tests
are being removed too.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since).

We should remove this code as it's no being used.