Remove live stream from content item helpers
The tests that used these helpers have been removed, so they are no
longer needed.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since).

We should remove this code as it's no being used.