Remove popular questions tests and helper methods
Popular questions was also part of the live stream section and should be
removed.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since).

We should remove this code as it's no being used.