Remove /transition routes
The taxon was renamed last week, so we no longer need this.

I've also moved the routes up, as this takes them above the call to DocumentTypeRoutingConstraint
in step by step routes. That constraint makes a content store request to work
out whether the route refers to a step by step page or not. It's best to
leave this out of the equation as much as possible.

https://trello.com/c/i4bVIMS0/1515-analytics-considerations-for-new-landing-page