Add non-JS version of timeline
- adds a form and radio buttons to select nation to show updates from
- copy is provisional and may change
- data isn't in yet, so displays demonstration messages

The button component has a margin_bottom of 30px. In the next commit the
button component will wrapped inside a show/hide details component and
it will look odd if there is a large margin between the end of the
button and the end of the details component, so the margin has been
removed completely.